splitted fft cpp into separate files (like the headers) added a cufft file and header based on FFTW
